Run web UI as a dedicated user
The web interface now runs as a dedicated "fusionforge" user, using ITK.
By giving that user read permissions to the *.ini files, including the
ones containing secrets, this removes the need for HTTP header trick to
securely pass secrets from config files to the PHP code.  This means
that we no longer need to regenerate an Apache config snippet and reload
Apache whenever a secret changes.