Change ownership of app data to the app user
The stored data (artifact attachments, docs, files and so on) are now
stored under a directory that belongs to the "fusionforge" user, rather
than the previous www-data/apache user.  This means that code running as
that www-data/apache user (for instance, any project website) no longer
has access to potentially restricted-access data.