Return effective time range for repository activities
Clients can specify a time range when requesting repository
activities, but it can get truncated if the server deems it to be too
long.  However, the client previously had no way to detect that.  So
this commit changes the return value of the
repositoryapi_repositoryActivity method to include the effective time
range that was used.